Create owensound.html
Code Breakdown
Head Section
The document is set to use UTF-8 encoding and a responsive viewport for mobile compatibility.
Leaflet.js CSS is included to style the map.
The #map div is styled to take up the entire viewport.
Body Section
Contains a div with the ID map, where the Leaflet.js map will be rendered.
Includes the Leaflet.js library from a CDN.
Initializes the map centered on Owen Sound with the coordinates [44.569, -80.940] and a zoom level of 13.
Adds the OpenStreetMap tile layer to the map.
JavaScript Section
Initializes the map and sets the view to Owen Sound's coordinates.
Defines a function addPin(lat, lng, popupText) to add markers to the map.
Listens for messages from Thunkable and adds markers based on the received data.
Usage
To use this HTML file in your Thunkable project, follow these steps:

Upload the map6.html file to your Backendless file storage.
Ensure the file is publicly accessible.
Set the URL of the Web Viewer component in Thunkable to point to the location of map6.html in Backendless. The URL should be in the format:
arduino
Copy code
https://aufaitapparel-us.backendless.app/files/map6.html
Use Thunkable blocks to send messages to the Web Viewer to dynamically add pins to the map.